Add the ability to include arbitrary files in the FAT Image
This PR adds the ability for users to include arbitrary files within the FAT image file created.
This functionality is controlled with the `--add_files <location_on_disk>:<location_within_image>` argument.
This flag can be supplied multiple times to include multiple files,
and the `<location_within_image>` can be excluded (eg: `--add_file <location_on_disk>`) to have the file
be placed at the root of the image with the name of the exterior file.