Create an empty uefi-run-lib package
This keeps the `uefi-run` app as the top-level package, but adds a
workspace member called `uefi-run-lib` in the `lib` subdirectory. Using
a separate package will make it easier to keep deps only needed by the
app (like `clap`) out of the library without relying on feature flags.

The `uefi-run` app now depends on `uefi-run-lib = "0.4"`, which of
course hasn't been released yet. So the dependency is locally patched to
point at the local directory. (This could be useful to keep even after
publishing the lib, so that dev/CI builds of the app are compiled
against the latest version of the lib.)